+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Synchronizer.getTransactionsByMemoSubstring()` has been added
 - `TransactionId` object has been added and used instead of `FirstClassByteArray` in `TransactionOverview` and 
   `PendingTransaction` model classes
+- `AccountUuid.toUuidString()` has been added, which prints a UUID value of `AccountUuid`
 
 ### Changed
 - `Synchronizer.getTransactions(accountUuid)` and `Synchronizer.transactions` now internally fill in